Add Regsiter() to register webservices and routes the path to GET, POST, and DELETE
Create a new WebService and add paths to it. A path is a URL endpoint and
routes are the path parameters or query parameters attached to the function
handlers. A webservice created to serve the Train resource. Three REST method
attached to three function handlers, getTrain, createTrain, and deleteTrain respectively
All these REST methods are defined on the instance of the TrainResource struct.